application/requirements.txt
asgiref==3.2.3
Django==2.2.9
django-annoying==0.10.5
django-appconf==1.0.3
django-imagekit==4.0.2
djangorestframework==3.10.3
gunicorn==20.0.4
pilkit==2.0
Pillow==6.2.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
virtualenv==16.7.9
whitenoise==4.1.4
